For bearing_String method --> changed parameter from float to double (since degrees are a double). -->also cleaned up Weather.m code. Don't need the WeatherDelegate.h anymore (didn't remove yet).

+++ Ship-Fit/Ship-Fit/Direction.h
@@ -16,7 +16,7 @@
 - (void)run_compass_withFilter: (CLLocationDegrees)compass_accuracy;
 - (void)halt_compass;
 - (void)set_bearing;
-+ (NSString*)bearing_String:(float)deg;
++ (NSString*)bearing_String:(double)deg;
 
 - (void)locationManager:(CLLocationManager *)manager
        didUpdateHeading:(CLHeading *)newHeading;
